Cleanup
Add GA_Agent class - seperate hyperparameter tuning and agent training
